eST-Miner adds basic test and file setup
To implement the eST-Miner, I use a test-driven-approach.
I added a unit test file that for now tries to apply the eST-Miner to the running-example.xes log.
As the algorithm has no implementation yet, this test fails. However this test already
illustrates how we want to expose the algorithm to users of the library and gives me
a starting point for further development iterations.