eST-Miner adds new test input
I added a new test input which contains an event log describing a Petri Net
with two long term dependencies and a XOR-choice.
These kinds of process structures are difficult to discover for other
process discovery algorithms, making it an ideal test candidate for verifying the
correctness of the eST-Miner implementation.