eST-Miner adds lexicographic order calculation
I added a test case and the corresponding functionality to calculate a
lexicographic ordering of the activities contained in an event log.
We require two orderings for traversing the search space of all possible places
while only requiring constant amount of memory to do so.
The orderings also contribute to how many places are checked to be fitting w.r.t the event log.
Hence, it is important for us to be able to calculate different kind of orderings
depending on the given event log. We can utilize the calculated orderings to speed up
the eST-Miner algorithm, if we are able to calculate orderings that result in a lot of pruning.

The added lexicographic ordering has no special properties, besides of being the ordering
used in the original paper by Mannel et al. (2019) when evaluating the eST-Miner.
In the paper it is also called a 'random ordering', as it ignores any properties that
might be extracted from the given event log.