eST-Miner reworks ordering unit tests
I splitted out test functions for testing ordering calculations based on log
metrics.
In addtion, we now use the running-example.xes event log to test the order calculations.
This is because this log gives us more variety in the contained traces and thus is more
interesting than the long_term_dependencies_xor.xes log when testing for correct order
calculations.