eST-Miner adds ordering based on average-first-occurrence-index
The average-first-occurrence-index (avgFOI) of an activity a, is the first index
in a trace where a is present, averaged over the number of traces where a occurrs.
If avgFOI(a) is low, we can expect the activity a to generate or consume a token early
on during replay of a trace. Places which have outgoing activities with low average occurrence
index are more likely to be underfed, as their output activities may require tokens early on
during replay, where none might be available.

I added a function and tests to calculate an ordering of activities in the log based on their
avgFOI.