eST-Miner refactors order calculation to output by ASC or DESC
To get variable input and output orderings, I want to be able to dynamically
control wether activities are ordered ascending or descending by the corresponding
metric.
For this I added a new OrderType enum that can be passed to each order calculating
function, telling it if the order should be ascending or descending.
I also added unit tests that verify that this works properly.
To improve readability of the test code, I extracted the unit tests for orderings into
their own class.