eST-Miner adds replay fitness calculation for places
Calculating the replay fitness properties (Fitting, Underfed, Overfed) for
a place is a fundamental part of the eST-Miner.
I added tests and functionality to solve this problem.
The corresponding function returns a fitness state that containas all
fitness properties for a given place.
We return a fitness state, because a place might be underfed and overfed
at the same time. If it is fitting, it cannot be either underfed nor overfed.